Change default cpu/memory resource limits for Controller Manager (#552)
For larger deployments, the default CPU limits of 500m (half a core) and
memory limit of 384Mi in the cluster might be insufficient. It has been
bumped up with this change and documentation has been added to alert the
user to modify these values in helm if they have larger clusters.

Co-authored-by: Nitish Bhat <nitish.bhat@amd.com>